Moved back-compat only functions out of media and into publicAPIs
